+++ packages/gamut-templates/src/index.ts
@@ -1,0 +1,2 @@
+export { default as Interstitial } from './Interstitial';
+export { default as SplitInterstitial } from './SplitInterstitial';