+++ packages/gamut/src/NotificationList/NotificationItem.tsx
@@ -4,7 +4,7 @@ import moment from 'moment';
 import Truncate from 'react-truncate';
 import { Notification } from './typings';
 import NotificationIcon from './NotificationIcon';
-import s from './styles/Notification.scss';
+import s from './styles/Notification.module.scss';
 
 export type NotificationItemProps = {
   onClick?: (event: object) => void;
@@ -15,7 +15,7 @@ const formatTime = (notificationDate: string): string => {
   return moment(notificationDate).fromNow();
 };
 
-export const NotificationItem = (props: NotificationItemProps) => {
+export const NotificationItem: React.FC<NotificationItemProps> = props => {
   const { notification, onClick } = props;
   const {
     date,