+++ packages/gamut/src/ToolTip/__tests__/ToolTip-test.tsx
@@ -3,12 +3,12 @@ import React from 'react';
 
 import { VisualTheme } from '../../theming/VisualTheme';
 import { ToolTip } from '..';
-import styles from '../styles.scss';
+import styles from '../styles.module.scss';
 
 describe('ToolTip', () => {
   it('adds the dark class to the container name when its theme is dark', () => {
     const wrapper = mount(
-      <ToolTip id="test-id" theme={VisualTheme.DarkMode}></ToolTip>
+      <ToolTip id="test-id" theme={VisualTheme.DarkMode} />
     );
 
     const containerClassName = wrapper
@@ -20,7 +20,7 @@ describe('ToolTip', () => {
 
   it('adds the light class to the container name when its theme is light', () => {
     const wrapper = mount(
-      <ToolTip id="test-id" theme={VisualTheme.LightMode}></ToolTip>
+      <ToolTip id="test-id" theme={VisualTheme.LightMode} />
     );
 
     const containerClassName = wrapper