+++ workspaces/gamut-storybook/stories/GamutIcons.stories.js
@@ -1,7 +1,8 @@
 import React from 'react';
-import { withKnobs, select, number } from '@storybook/addon-knobs';
+import { withKnobs, select, number, boolean } from '@storybook/addon-knobs';
 import * as icons from '@codecademy/gamut-icons';
 import { selectableColors } from './helpers';
+import cx from 'classnames';
 import s from './Icon-story.scss';
 
 export default {
@@ -13,11 +14,17 @@ export const allIcons = () => (
   <div className={s.grid}>
     {Object.entries(icons).map(([iconName, Icon]) => (
       <span className={s.iconWrapper} key={iconName}>
-        <Icon
-          key={iconName}
-          size={number('size', 64)}
-          color={select('color', selectableColors, selectableColors.black)}
-        />
+        <span
+          className={cx({
+            [s.iconGrid]: boolean('show grid', false),
+          })}
+        >
+          <Icon
+            key={iconName}
+            size={number('size', 64)}
+            color={select('color', selectableColors, selectableColors.black)}
+          />
+        </span>
         <span className={s.iconLabel}>{iconName}</span>
       </span>
     ))}